assetData + dynamically generate earned data
