+++ pages/api/llama/[address].ts
@@ -6,6 +6,12 @@ import { IncomingLlamaUserData, LlamaTier, TierToLevel } from 'types/llama'
 
 import { validateLlamaPfpAllowList } from 'api/premintCheck/checks/llama'
 
+import { fetcher } from 'utils/requests'
+
+import { allRows } from './assetData'
+
+const token = ''
+
 const getMaxTier = (tiersBySeason: Record<number, LlamaTier>) => {
     const tiers = Object.values(tiersBySeason)
     const maxTier = tiers.reduce((acc: string, curr: string) => {
@@ -16,54 +22,6 @@ const getMaxTier = (tiersBySeason: Record<number, LlamaTier>) => {
     }, 'Traveler') as string
     return maxTier
 }
-
-const llamaUserData: IncomingLlamaUserData = {
-    id: 92,
-    created_at: '2022-08-12T19:06:20.378726+00:00',
-    username: 'garcia',
-    // email: 'jerry@llama.xyz',
-    tier: LlamaTier.Traveler,
-    tier_by_season: {
-        1: null,
-        2: null,
-        3: null,
-        4: LlamaTier.Traveler,
-    },
-    eth_login_address: '0xeB9f74528aA4F9cA15612c68dC3eE944dF808611',
-    admin: true,
-}
-
-const llamaBackgroundSparklyRow: LayerItemRow = {
-    project: 'llama',
-    category: 'background',
-    modifiable: true,
-    name: 'Sparkly',
-    pngLink: 'https://llama-assets.s3.amazonaws.com/backgrounds/sparkly.png',
-    earnedDescription: 'Earned by being a Traveler',
-    earnedCriteria: 'Traveler',
-}
-
-const llamaBackgroundStarryRow: LayerItemRow = {
-    project: 'llama',
-    category: 'background',
-    modifiable: true,
-    name: 'Starry Night',
-    pngLink: 'https://llama-assets.s3.amazonaws.com/backgrounds/starry.png',
-    earnedDescription: 'Earned by being a Explorer',
-    earnedCriteria: 'Explorer',
-}
-
-const llamaBodyBrownRow: LayerItemRow = {
-    project: 'llama',
-    category: 'body',
-    modifiable: false,
-    name: 'Brown',
-    pngLink: 'https://llama-assets.s3.amazonaws.com/body/brown.png',
-    earnedDescription: 'available to all',
-    earnedCriteria: null,
-}
-
-const layersArr = [llamaBackgroundSparklyRow, llamaBackgroundStarryRow, llamaBodyBrownRow]
 
 const llamaBackgroundCriteriaChecker = (criteriaRequirement: string) => (llamaUserData: IncomingLlamaUserData) => {
     const llamaLevel = TierToLevel[llamaUserData.tier]
@@ -79,13 +37,13 @@ const llamaNecklaceCriteriaChecker = (criteriaRequirement: string) => (llamaUser
 }
 
 const llamaCriteriaMap: CriteriaMap = {
-    background: {
+    Background: {
         Traveler: llamaBackgroundCriteriaChecker(LlamaTier.Traveler),
         Explorer: llamaBackgroundCriteriaChecker(LlamaTier.Explorer),
         Mountaineer: llamaBackgroundCriteriaChecker(LlamaTier.Mountaineer),
         Rancher: llamaBackgroundCriteriaChecker(LlamaTier.Rancher),
     },
-    necklace: {
+    Necklace: {
         Traveler: llamaNecklaceCriteriaChecker(LlamaTier.Traveler),
         Explorer: llamaNecklaceCriteriaChecker(LlamaTier.Explorer),
         Mountaineer: llamaNecklaceCriteriaChecker(LlamaTier.Mountaineer),
@@ -100,70 +58,23 @@ const layerItemRowToLayerItem = (
 ): LayerItemData => {
     const { name, pngLink, earnedDescription, earnedCriteria, category } = layerItemRow
 
+    // if (category === 'Background') {
+    //     console.log(earnedCriteria)
+    //     console.log(criteriaMap?.[category]?.[earnedCriteria]?.(userData))
+    // }
     return {
         name,
         pngLink,
         earnedDescription,
-        earned: criteriaMap?.[category]?.[earnedCriteria]?.(userData) || null,
+        earned: criteriaMap?.[category]?.[earnedCriteria]?.(userData) ?? null,
     }
 }
-
-const userData: UserData = {
-    minted: true,
-    tokenId: 12,
-    nftName: `austingreen.eth's Llama`,
-    currentLayers: [
-        { category: 'Background', name: 'Sparkly' },
-        { category: 'Necklace', name: 'Green' },
-    ],
-}
-
-const assetData: AssetData = [
-    {
-        category: 'Background',
-        modifiable: true,
-        options: [
-            {
-                name: 'Sparkly',
-                pngLink: 'https://llama-assets.s3.amazonaws.com/background/sparkly.png',
-                earned: true,
-                earnedDescription: 'Reach Traveler Tier',
-            },
-            {
-                name: 'Starry Night',
-                pngLink: 'https://llama-assets.s3.amazonaws.com/background/starry_night.png',
-                earned: false,
-                earnedDescription: 'Reach Explorer Tier',
-            },
-        ],
-    },
-    {
-        category: 'Necklace',
-        modifiable: true,
-        options: [
-            {
-                name: 'Green',
-                pngLink: 'https://llama-assets.s3.amazonaws.com/necklace/green.png',
-                earned: true,
-                earnedDescription: 'Reach Traveler Tier',
-            },
-            {
-                name: 'Blue',
-                pngLink: 'https://llama-assets.s3.amazonaws.com/necklace/blue.png',
-                earned: false,
-                earnedDescription: 'Reach Explorer Tier',
-            },
-        ],
-    },
-]
 
 const layerItemRowsToAssetData = (
     layerRows: LayerItemRow[],
     userData: IncomingLlamaUserData,
     criteriaMap: CriteriaMap,
 ): AssetData => {
-    // get all layer items for a project
-    // group by category
     // loop through them all to get the earned status of each
 
     // group by category
@@ -176,24 +87,27 @@ const layerItemRowsToAssetData = (
         options: val.map((obj) => layerItemRowToLayerItem(obj, userData, criteriaMap)),
     }))
 
-    // console.log(layerRowsByCategoryMap);
-    console.log(assetData)
-
     return assetData.value()
 }
 
-const testData = layerItemRowsToAssetData(layersArr, llamaUserData, llamaCriteriaMap)
-
-console.log(testData)
+const llamaUsersBaseUrl = 'https://community.llama.xyz/api/users/'
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
     try {
         const address = req.query.address as string
-        const jwt = req.query.check as string
+        const authToken = req.query.check || (token as string)
+        const llamaUserId = req.query.llamaUserId as string
 
-        const checkResponse = validateLlamaPfpAllowList(address, jwt)
+        const llamaData = await fetcher(llamaUsersBaseUrl + llamaUserId, {
+            headers: { authorization: `Bearer ${token}` },
+        })
 
-        res.status(200).json(address)
+        const testData = layerItemRowsToAssetData(allRows, llamaData.body, llamaCriteriaMap)
+        // const checkResponse = await validateLlamaPfpAllowList(address, jwt)
+
+        console.log('llama data', llamaData)
+
+        res.status(200).json(testData)
     } catch (error) {
         console.log(error)
         res.status(500).json({ error: error.message })
