+++ pages/api/premintCheck/checks/llama.ts
@@ -1,9 +1,5 @@
 import { CheckResponse, createDomainSeparator, generateSignature } from 'utils/premint'
 
-export const llamaS3BucketName = 'llama-pfp'
-
-// https://llama-pfp.s3.us-east-1.amazonaws.com/Layers/Background/Blue_-_Pink_-_Orange.png
-
 const llamaPfpDomainSeparator = createDomainSeparator('Llama PFP', '0x17a059b6b0c8af433032d554b0392995155452e6') // TODO add a contract address
 
 export const validateLlamaPfpAllowList = async (address: string, llamaJwt: string): Promise<CheckResponse> => {
